Change the way web pages rankings are calculated
It was decided to refuse to rank pages that have not yet
been crawled. Now this causes problems with rating
calculation, as the PR algorithm does not work correctly
with trees, namely, the sum of the ratings of each page is
not equal to 1. The problem is solved by analysing more
web pages and links. The more data examined, the more
relationships between pages are found and, therefore, the
graph becomes less like a tree

The reason for the decision was poor visualization for
partial data. When the user loads 10 web pages for
analysis, he expects to see the relationship between them,
excluding third-party pages