Change the parsing of emails
Now emails are parsed in a slightly different way, focused
on further visualization. Data is not immediately saved to
the database, but temporarily stored in RAM in the form of
a records list. So I avoid excessive decomposition of the
project and the size of the database(s), since in the
original project the data is first loaded into the
database, then control is transferred to another file that
processes them and stores them in another database. I think
this approach is not optimal