Change the database structure
The database structure has been changed to reduce memory
consumption. The recipient of emails as well as names are
no longer stored in the database. For future visualization,
this data is redundant.

Also, server names (domain names of mail servers) are now
stored in a separate table to save memory